+++ README.md
@@ -5,6 +5,18 @@
 - цена (price) - доступ ограничен
 - количество в наличии (quantity)
 
+## Наследники класса Product
+# Smartphone
+- производительность (efficiency)
+- модель (model)
+- объем памяти (memory)
+- цвет (color)
+
+# LawnGrass
+- cnhfyf ghjbpdjlbntkm (counry)
+- срок проростания (germenation_period)
+- цвет (color)
+
 ### Метод new_product
 Преобразует словарь в объект класса
 
@@ -17,6 +29,7 @@
 
 ### __add__
 Суммирует цену запрашиваемых товаров, оставшихся на складе
+Невозможно сложение объектов разных классов
 
 
 ## Класс Category
@@ -29,6 +42,7 @@
 
 ### Метод add_product
 Добавляет новый продукт в список
+Исключена возможность добавления объекта не являющегося продуктом или его наследником
 
 ### __str__
 Выводит читабельную информацию о категории
