+++ main.py
@@ -1,138 +1,83 @@
-class Product:
-    """ Класс для работы с информацией о продукте """
+from src.catalog import Category
+from src.product import LawnGrass, Product, Smartphone
 
-    name: str
-    description: str
-    price: float
-    quantity: int
-
-    def __init__(self, name, description, price, quantity):
-        """ Функция инициализации продукта"""
-
-        self.name = name
-        self.description = description
-        self.__price = price
-        self.quantity = quantity
-
-    @classmethod
-    def new_product(cls, new_element):
-        """ Создает объект класса из словаря """
-
-        name = new_element["name"]
-        description = new_element["description"]
-        price = new_element["price"]
-        quantity = new_element["quantity"]
-        return cls(name, description, price, quantity)
-
-    @property
-    def price(self):
-        """ Возвращает приватный атрибут "price" """
-
-        return self.__price
-
-    @price.setter
-    def price(self, cost):
-        """ Проверяет "price" на положительность """
-
-        if cost <= 0:
-            print("Цена не должна быть нулевая или отрицательная")
-
-        else:
-            self.__price = cost
-
-    def __str__(self):
-        """ Выводит строку с именем, ценой и колличеством продукта """
-
-        return f"{self.name}, {self.price} руб. Остаток: {self.quantity} шт."
-
-    def __add__(self, other):
-        """ Возвращает сумму продуктов """
-
-        return self.__price * self.quantity + other.__price * other.quantity
-
-
-class Category:
-    """ Класс для работы с категориями """
-
-    category_count = 0
-    product_count = 0
-
-    name: str
-    description: str
-    __products: list
-
-    def __init__(self, name, description, products):
-        """ Функция инициализации категории """
-
-        self.name = name
-        self.description = description
-        self.__products = products
-        Category.product_count += len(products)
-        Category.category_count += 1
-
-    def add_product(self, product: Product):
-        """ Метод добавляющий новый продукт """
-
-        if isinstance(product, Product):
-            self.__products.append(product)
-            Category.product_count += 1
-
-        else:
-            print("Не является объектом класса \"Product\"")
-
-    @property
-    def products(self):
-
-        """ Возвращает строку с описанием всех продуктов в категории """
-        product_str = ""
-        for element in self.__products:
-            product_str += \
-                f"""{element.name}, {element.description}, {
-                element.price} руб. Остаток: {element.quantity} шт.\n"""
-        return product_str
-
-    def __str__(self):
-        """ Возвращает имя категории и колличество продуктов """
-
-        amount = 0
-        for element in self.__products:
-            amount += element.quantity
-
-        return f"{self.name}, колличество продуктов: {amount}"
-
-
-if __name__ == '__main__':
-    product1 = Product(
-        "Samsung Galaxy S23 Ultra",
-        "256GB, Серый цвет, 200MP камера",
-        180000.0,
-        5)
-    product2 = Product(
-        "Iphone 15",
-        "512GB, Gray space",
-        210000.0,
-        8)
-    product3 = Product(
-        "Xiaomi Redmi Note 11",
-        "1024GB, Синий",
-        31000.0,
-        14)
-
-    print(str(product1))
-    print(str(product2))
-    print(str(product3))
-
-    category1 = Category(
-        "Смартфоны",
-        """Смартфоны, как средство не только коммуникации,"
-         но и получения дополнительных функций для удобства жизни""",
-        [product1, product2, product3]
-    )
-
-    print(str(category1))
-
-    print(category1.products)
-
-    print(product1 + product2)
-    print(product1 + product3)
-    print(product2 + product3)
+# if __name__ == '__main__':
+#     smartphone1 = Smartphone("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5, 95.5,
+#                          "S23 Ultra", 256, "Серый")
+#     smartphone2 = Smartphone("Iphone 15", "512GB, Gray space", 210000.0, 8, 98.2, "15", 512, "Gray space")
+#     smartphone3 = Smartphone("Xiaomi Redmi Note 11", "1024GB, Синий", 31000.0, 14, 90.3, "Note 11", 1024, "Синий")
+#
+#     print(smartphone1.name)
+#     print(smartphone1.description)
+#     print(smartphone1.price)
+#     print(smartphone1.quantity)
+#     print(smartphone1.efficiency)
+#     print(smartphone1.model)
+#     print(smartphone1.memory)
+#     print(smartphone1.color)
+#
+#     print(smartphone2.name)
+#     print(smartphone2.description)
+#     print(smartphone2.price)
+#     print(smartphone2.quantity)
+#     print(smartphone2.efficiency)
+#     print(smartphone2.model)
+#     print(smartphone2.memory)
+#     print(smartphone2.color)
+#
+#     print(smartphone3.name)
+#     print(smartphone3.description)
+#     print(smartphone3.price)
+#     print(smartphone3.quantity)
+#     print(smartphone3.efficiency)
+#     print(smartphone3.model)
+#     print(smartphone3.memory)
+#     print(smartphone3.color)
+#
+#     grass1 = LawnGrass("Газонная трава", "Элитная трава для газона", 500.0, 20, "Россия", "7 дней", "Зеленый")
+#     grass2 = LawnGrass("Газонная трава 2", "Выносливая трава", 450.0, 15, "США", "5 дней", "Темно-зеленый")
+#
+#     print(grass1.name)
+#     print(grass1.description)
+#     print(grass1.price)
+#     print(grass1.quantity)
+#     print(grass1.country)
+#     print(grass1.germination_period)
+#     print(grass1.color)
+#
+#     print(grass2.name)
+#     print(grass2.description)
+#     print(grass2.price)
+#     print(grass2.quantity)
+#     print(grass2.country)
+#     print(grass2.germination_period)
+#     print(grass2.color)
+#
+#     smartphone_sum = smartphone1 + smartphone2
+#     print(smartphone_sum)
+#
+#     grass_sum = grass1 + grass2
+#     print(grass_sum)
+#
+#     try:
+#         invalid_sum = smartphone1 + grass1
+#     except TypeError:
+#         print("Возникла ошибка TypeError при попытке сложения")
+#     else:
+#         print("Не возникла ошибка TypeError при попытке сложения")
+#
+#     category_smartphones = Category("Смартфоны", "Высокотехнологичные смартфоны", [smartphone1, smartphone2])
+#     category_grass = Category("Газонная трава", "Различные виды газонной травы", [grass1, grass2])
+#
+#     category_smartphones.add_product(smartphone3)
+#
+#     print(category_smartphones.products)
+#
+#     print(Category.product_count)
+#
+#     try:
+#         category_smartphones.add_product("Not a product")
+#     except TypeError:
+#         print("Возникла ошибка TypeError при добавлении не продукта")
+#     else:
+#         print("Не возникла ошибка TypeError при добавлении не продукта")
